refactor(admin-tx): use enum for supported tx type

+++ tests/harness/admin.rs
@@ -138,7 +138,7 @@ impl AdminContext {
         let sig_set = create_signature_set(&self.privkeys, &self.signer_indices, sighash);
         let payload = borsh::to_vec(&SignedPayload::new(seqno, action.clone(), sig_set))
             .expect("serialization should succeed");
-        (payload, action.tx_type())
+        (payload, action.tx_type() as u8)
     }
 }
